requirements.txt
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-extra-fields==3.7.0
PyJWT==2.8.0
python-dotenv==1.0.0
djoser==2.2.0
drf-nested-routers==0.93.4
drf-yasg==1.21.7
requests==2.31.0
gunicorn==21.2.0
Pillow==10.0.1
django-debug-toolbar==4.2.0